Move view list related components to views folder
